refactor(web): reorganize auth/products into feature-first layers
