+++ apps/web/src/app/app.routes.ts
@@ -12,13 +12,13 @@ export const routes: Routes = [
     path: 'login',
     canActivate: [guestGuard],
     loadComponent: () =>
-      import('./features/auth/pages/login.page').then((m) => m.LoginPageComponent),
+      import('./features/auth/ui/pages/login.page').then((m) => m.LoginPageComponent),
   },
   {
     path: 'register',
     canActivate: [guestGuard],
     loadComponent: () =>
-      import('./features/auth/pages/register.page').then(
+      import('./features/auth/ui/pages/register.page').then(
         (m) => m.RegisterPageComponent,
       ),
   },
@@ -26,7 +26,7 @@ export const routes: Routes = [
     path: 'products',
     canActivate: [authGuard],
     loadComponent: () =>
-      import('./features/products/pages/products.page').then(
+      import('./features/products/ui/pages/products.page').then(
         (m) => m.ProductsPageComponent,
       ),
   },
