+++ apps/web/src/app/shared/README.md
@@ -1,0 +1,8 @@
+# Shared Layer
+
+This folder contains reusable UI building blocks, pipes, directives, and pure utilities shared across features.
+
+Rules:
+- No feature-specific business logic.
+- No direct API calls.
+- Keep artifacts framework-agnostic when possible.
